audio string pool works
