+++ src/game/player-ship.ts
@@ -23,7 +23,7 @@ import { BulletDef } from "./bullet.js";
 import { DeletedDef } from "../delete.js";
 import { clamp, min } from "../math.js";
 import { createCannon } from "./cannon.js";
-import { MusicDef } from "../music.js";
+import { AudioDef } from "../audio.js";
 import { LocalPlayerDef, PlayerDef } from "./player.js";
 import { CameraDef } from "../camera.js";
 import { InputsDef } from "../inputs.js";
@@ -324,7 +324,7 @@ export function registerShipSystems(em: EntityManager) {
     "ship-hit",
     [[PlayerShipLocalDef]] as const,
     ([ship], partIdx: number) => {
-      const music = em.getResource(MusicDef)!;
+      const music = em.getResource(AudioDef)!;
       const part = ship.playerShipLocal.parts[partIdx]()!;
       part.renderable.enabled = false;
       part.shipPart.damaged = true;
@@ -340,7 +340,7 @@ export function registerShipSystems(em: EntityManager) {
   em.registerSystem(
     [PlayerShipPropsDef, PlayerShipLocalDef, PositionDef, AuthorityDef],
     [
-      MusicDef,
+      AudioDef,
       InputsDef,
       CameraDef,
       GameStateDef,
